Fixed issue regarding remaining view in Papyrus model set.
Papyrus stores information about the pages opened in the editor. When
opening the editor for the same model, the information is read. The
file, however, contains absolute CDO uris including the checkout number.
If this number changes, the pages cannot be opened.

This leads to broken editor pages AND a wrong value in the view property
of the model set. This prohibits cleaning up properly.